+++ OreVeins/src/defaultPackadge/Gold.java
@@ -60,7 +60,7 @@ public class Gold
 		}
 		else
 		{
-			probToSpawn = config.getInt(path+".SpawnProbability");
+			probToSpawn = config.getDouble(path+".SpawnProbability");
 		}
 	}
 }
