+++ OreVeins/src/defaultPackadge/Iron.java
@@ -61,7 +61,7 @@ public class Iron
 		}
 		else
 		{
-			probToSpawn = config.getInt(path+".SpawnProbability");
+			probToSpawn = config.getDouble(path+".SpawnProbability");
 		}
 	}
