+++ OreVeins/src/defaultPackadge/Redstone.java
@@ -61,7 +61,7 @@ public class Redstone
 		}
 		else
 		{
-			probToSpawn = config.getInt(path+".SpawnProbability");
+			probToSpawn = config.getDouble(path+".SpawnProbability");
 		}
 		
 	}
