+++ OreVeins/src/defaultPackadge/TertiaryVein.java
@@ -99,7 +99,7 @@ public class TertiaryVein {
 		}
 		else
 		{
-			probToSpawn = config.getInt(string+".SpawnProbability");
+			probToSpawn = config.getDouble(string+".SpawnProbability");
 		}
 	}
